Update: to the version 2.0 (40%)
