+++ src/theme/Icon/Edit/index.js
