+++ src/theme/NavbarItem/DocsVersionNavbarItem.tsx
@@ -1,0 +1,27 @@
+/**
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+import React from 'react';
+import {useDocsVersionCandidates} from '@docusaurus/theme-common/internal';
+import DefaultNavbarItem from '@theme/NavbarItem/DefaultNavbarItem';
+import type {Props} from '@theme/NavbarItem/DocsVersionNavbarItem';
+import type {GlobalVersion} from '@docusaurus/plugin-content-docs/client';
+
+const getVersionMainDoc = (version: GlobalVersion) =>
+  version.docs.find((doc) => doc.id === version.mainDocId)!;
+
+export default function DocsVersionNavbarItem({
+  label: staticLabel,
+  to: staticTo,
+  docsPluginId,
+  ...props
+}: Props): JSX.Element {
+  const version = useDocsVersionCandidates(docsPluginId)[0];
+  const label = staticLabel ?? version.label;
+  const path = staticTo ?? getVersionMainDoc(version).path;
+  return <DefaultNavbarItem {...props} label={label} to={path} />;
+}
