+++ src/theme/NavbarItem/LocaleDropdownNavbarItem/styles.module.css
