+++ src/theme/NavbarItem/SearchNavbarItem.tsx
@@ -1,0 +1,26 @@
+/**
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+import React from 'react';
+import SearchBar from '@theme/SearchBar';
+import NavbarSearch from '@theme/Navbar/Search';
+import type {Props} from '@theme/NavbarItem/SearchNavbarItem';
+
+export default function SearchNavbarItem({
+  mobile,
+  className,
+}: Props): JSX.Element | null {
+  if (mobile) {
+    return null;
+  }
+
+  return (
+    <NavbarSearch className={className}>
+      <SearchBar />
+    </NavbarSearch>
+  );
+}
