--- src/theme_old/Card/index.jsx
@@ -1,2 +1,0 @@
-export * from './Card'
-export * from './CardGradient'
