--- src/theme_old/CardList/styles.module.css
@@ -1,171 +1,0 @@
-.cardImg {
-  width: 190px;
-  height: 190px;
-  margin: 0 auto;
-  flex-shrink: 0;
-  background-position: center center !important;
-  background-size: contain !important;
-  background-repeat: no-repeat !important;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.cardImg > div {
-  font-size: 1.5rem;
-  font-weight: 600;
-  max-width: 50px;
-  line-height: 1.3;
-  text-align: center;
-  color: var(--ifm-color-light);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
-}
-
-.cardImg > div span {
-  font-size: 1rem;
-  display: block;
-}
-
-.cardInner {
-  padding: 2rem;
-  display: flex;
-  height: 100%;
-  flex-direction: column;
-  justify-content: space-between;
-}
-
-.cardContainer {
-  padding-top: 20px;
-  border-radius: 2px;
-  color: var(--ifm-color-emphasis-800);
-  --ifm-link-color: var(--ifm-color-emphasis-800);
-  --ifm-link-hover-color: var(--ifm-color-emphasis-800);
-  --ifm-link-hover-decoration: none;
-
-  box-shadow: none;
-  border: 2px solid var(--ifm-color-emphasis-200);
-  transition: box-shadow var(--ifm-transition-slow) ease, background-color var(--ifm-transition-fast) ease;
-  margin-bottom: 0;
-}
-
-.cardContainer:hover a{
-  text-decoration: none;
-}
-
-.cardContainer.cardContainerLink:hover {
-  /* box-shadow: var(--ifm-global-shadow-md); */
-  box-shadow: 0 4px 16px 0 rgb(0 0 0 / 10%);
-}
-
-[data-theme="dark"] .cardContainer.cardContainerLink:hover {
-  --ifm-card-background-color: #2d2d2d; /* original, non-hovered color is #242526 */
-}
-
-.cardContainer:not(.cardContainerLink) {
-  cursor: not-allowed;
-}
-
-.cardTitle {
-  font-size: 1rem;
-  min-height: 1.2rem;
-  margin-bottom: 0;
-  font-weight: 600;
-  text-align: center;
-}
-
-.cardDescription {
-  font-size: 0.8rem;
-}
-
-a.cardLearnMore {
-  font-size: 1rem;
-  text-align: center;
-  text-decoration: underline;
-  margin-top: 20px;
-}
-
-.filterSelects {
-  width: calc((100% - 100px) + 1rem);
-  margin-left: -0.5rem;
-  flex-shrink: 0;
-  display: flex;
-  align-items: flex-start;
-}
-
-.filterSelects > div {
-  padding: 0 0.5rem;
-  width: 100%;
-}
-
-.filter {
-  display: flex;
-  align-items: stretch;
-}
-
-.filterLabel {
-  width: 100px;
-  padding-top: 6px;
-  display: flex;
-  flex-shrink: 0;
-  padding-right: 1rem;
-  align-items: flex-start;
-}
-
-.notFound {
-  display: block;
-  text-align: center;
-}
-
-
-@media (max-width: 1400px) {
-  .cardImg {
-    width: 140px;
-    height: 140px;
-  }
-
-  .cardImg > div {
-    font-size: 1.125rem;
-  }
-  
-  .cardImg > div span {
-    font-size: 0.9375rem;
-  }
-}
-
-@media (max-width: 1200px) {
-  .cardImg {
-    width: 120px;
-    height: 120px;
-  }
-
-  .cardImg > div {
-    font-size: 1rem;
-  }
-  
-  .cardImg > div span {
-    font-size: 0.875rem;
-  }
-}
-
-@media (max-width: 997px) {
-  .filter {
-    flex-direction: column;
-  }
-
-  .filterLabel {
-    margin-bottom: 0.5rem;
-  }
-
-  .filterSelects {
-    flex-direction: column;
-  }
-
-  .filterSelects > div {
-    margin-bottom: 0.5rem;
-    width: 100%;
-  }
-}
-
