--- src/theme_old/EditThisPage/index.js
@@ -1,28 +1,0 @@
-/**
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-import React from 'react';
-import Translate from '@docusaurus/Translate';
-import IconEdit from '@theme/IconEdit';
-import styles from './styles.module.css';
-
-export default function EditThisPage({ editUrl }) {
-  return (
-    <a
-      href={editUrl}
-      target="_blank"
-      rel="noreferrer noopener"
-      className={styles.editLabel}>
-      <IconEdit />
-      <Translate
-        id="theme.common.editThisPage"
-        description="The link label to edit the current page"
-      >
-        Edit this page
-      </Translate>
-    </a>
-  );
-}
