--- src/theme_old/Hero/styles.module.css
@@ -1,21 +1,0 @@
-.hero {
-  text-align: center;
-  margin-bottom: 2.6rem;
-}
-
-.hero img{
-  max-width: 250px;
-  max-height: 250px;
-  margin-bottom: 1rem;
-  filter: var(--ifm-menu-link-sublist-icon-filter);
-}
-
-.hero h1 {
-  font-size: 1.875rem;
-  margin-bottom: 1.6rem;
-  font-family: var(--ifm-font-family-secondary);
-}
-
-.hero p {
-  font-size: 1rem;
-}
